SWARM-1542 Run Cassandra,MongoDB,Neo4j,OrientDB test with docker via 'mvn -Pdocker clean install', added 120 second timeout, renamed -pnosql in root pom to -pdocker, use mapped for MongoDB, log scanning for Cassandra readiness
