+++ testsuite/testsuite-cassandra/src/it/java/org/wildfly/swarm/cassandra/test/StatefulTestBean.java
