+++ testsuite/testsuite-cassandra/src/it/resources/META-INF/beans.xml
