+++ testsuite/testsuite-mongodb/src/it/resources/WEB-INF/beans.xml
