+++ testsuite/testsuite-neo4j/pom.xml
@@ -65,4 +65,77 @@
     </dependency>
   </dependencies>
 
+  <profiles>
+    <profile>
+      <id>docker</id>
+      <build>
+        <plugins>
+          <plugin>
+            <artifactId>maven-failsafe-plugin</artifactId>
+            <configuration>
+              <includes>
+                <include>**/*IT.java</include>
+              </includes>
+              <systemPropertyVariables>
+                <org.apache.maven.user-settings>${session.request.userSettingsFile.path}</org.apache.maven.user-settings>
+              </systemPropertyVariables>
+            </configuration>
+          </plugin>
+          
+          <plugin>
+            <groupId>io.fabric8</groupId>
+            <artifactId>docker-maven-plugin</artifactId>
+            <version>0.22.1</version>
+            <configuration>
+              <images>
+                <image>
+                  <name>neo4j</name>
+                  <run>
+                    <env>
+                        <NEO4J_AUTH>neo4j/changethis</NEO4J_AUTH>
+                    </env>
+                    <privileged>true</privileged>
+                    <ports>
+                      <port>7687:7687</port>
+                      <port>7474:7474</port>
+                    </ports>
+                    <wait>
+                        <time>120000</time>
+                        <tcp>
+                            <ports>
+                                <port>7687</port>
+                                <port>7474</port>
+                            </ports>
+                        </tcp>
+                    </wait>
+                    <log>
+                    <prefix>TC</prefix>
+                        <date>default</date>
+                        <color>cyan</color>
+                    </log>
+                  </run>
+                </image>
+              </images>
+            </configuration>
+              <executions>
+                <execution>
+                  <id>pre-integration-test</id>
+                  <phase>pre-integration-test</phase>
+                  <goals>
+                    <goal>start</goal>
+                  </goals>
+                </execution>                        
+                <execution>
+                  <id>post-integration-test</id>
+                  <phase>post-integration-test</phase>
+                  <goals>
+                    <goal>stop</goal>
+                  </goals>
+                </execution>
+              </executions>
+          </plugin>
+        </plugins>
+      </build>
+    </profile>
+  </profiles>
 </project>
