+++ testsuite/testsuite-neo4j/src/it/java/org/wildfly/swarm/neo4j/test/NestedBean.java
