+++ testsuite/testsuite-neo4j/src/it/resources/META-INF/beans.xml
