+++ testsuite/testsuite-orientdb/src/it/resources/META-INF/beans.xml
